backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.2.1

# MongoDB
motor==3.6.0
pymongo==4.9.1
pydantic==2.12.5
pydantic-settings==2.7.1
email-validator==2.1.0

# Firebase Admin SDK
firebase-admin==6.6.0

# ML Dependencies (from clinical_ml - COPY AS-IS)
torch==2.9.1
transformers==4.57.3
sentence-transformers==5.2.0
faiss-cpu==1.13.2
openai==2.14.0
pdfminer.six==20251230
pytesseract==0.3.13
pillow==12.1.0
reportlab==4.4.7
numpy==2.2.6
scikit-learn==1.7.2
regex==2025.11.3
requests==2.32.5
pyyaml==6.0.3
joblib==1.5.3
scipy==1.15.3
